CAF - Working audio data reading

+++ ATL/AudioData/IO/ID3v2.cs
@@ -1333,7 +1333,7 @@ namespace ATL.AudioData.IO
 
             if (tagHeader.UsesUnsynchronisation)
             {
-                MemoryStream s = new MemoryStream(Size);
+                MemoryStream s = new MemoryStream((int)Size);
                 using (BinaryWriter w = new BinaryWriter(s, tagEncoding))
                 {
                     result = writeChaptersInternal(writer, w, chapters, tagEncoding, writer.BaseStream.Position);
@@ -1552,7 +1552,7 @@ namespace ATL.AudioData.IO
 
             if (tagHeader.UsesUnsynchronisation && !isInsideUnsynch)
             {
-                s = new MemoryStream(Size);
+                s = new MemoryStream((int)Size);
                 w = new BinaryWriter(s, tagEncoding);
                 frameOffset = writer.BaseStream.Position;
             }
@@ -1721,7 +1721,7 @@ namespace ATL.AudioData.IO
 
             if (tagHeader.UsesUnsynchronisation && !isInsideUnsynch)
             {
-                s = new MemoryStream(Size);
+                s = new MemoryStream((int)Size);
                 w = new BinaryWriter(s, usedTagEncoding);
                 frameOffset = writer.BaseStream.Position;
             }
